--story=117513038 feat：部署节点支持手动安装 GSE Agent
1.重构配置属性读取代码；
2.抽离Agent安装任务管理到独立的类；
3.重命名GSE Agent操作相关类与包，与蓝盾Agent区分。

+++ src/backend/ci/ext/tencent/environment/biz-environment-tencent/src/main/kotlin/com/tencent/devops/environment/pojo/job/ccreq/CCFindHostBizRelationsReq.kt
@@ -18,5 +18,5 @@ data class CCFindHostBizRelationsReq(
     val bkUsername: String? = "admin",
     @get:Schema(title = "主机属性列表", required = true)
     @JsonProperty("bk_host_id")
-    val bkHostId: List<Long>
+    val bkHostId: List<Int>
 )
